% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthomr-package.R
\docType{package}
\name{orthomr-package}
\alias{orthomr}
\alias{orthomr-package}
\title{orthomr: Orthogonal Polynomials for Non-Linear Mendelian Randomization}
\description{
More to come.
}
\author{
\strong{Maintainer}: Guillaume Butler-Laporte \email{guillaume.butler-laporte@mcgill.ca}

}
\keyword{internal}
