% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mvmr.R
\name{predict_mvmr}
\alias{predict_mvmr}
\title{Predict Using MVMR Polynomial Coefficients}
\usage{
predict_mvmr(exposure_values, polynomial_coefficients)
}
\arguments{
\item{exposure_values}{Numeric vector. Exposure values.}

\item{polynomial_coefficients}{List. Polynomial coefficients from \code{obtain_final_coeffs()}.}
}
\value{
Numeric vector of predicted outcome values.
}
\description{
Given exposure values and polynomial coefficients from MVMR, predicts
the outcome values.
}
\examples{
\dontrun{
exposure_new <- seq(-3, 3, 0.1)
outcome_predicted <- predict_mvmr(exposure_new, final_coefficients)
}
}
