% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_exp_inst.R
\name{merge_exp_inst}
\alias{merge_exp_inst}
\title{Merge Exposure Instruments}
\usage{
merge_exp_inst(gwas_exp)
}
\arguments{
\item{gwas_exp}{List of data frames. Each data frame should contain columns:
ins, beta, se, pval, af.}
}
\value{
A data frame with merged instruments in TwoSampleMR format with columns:
SNP, exposure, id.exposure, effect_allele.exposure, other_allele.exposure,
beta.exposure, se.exposure, pval.exposure, eaf.exposure.
}
\description{
Loops through GWAS exposure list, extracts significant instruments from all
GWAS, merges them, and removes duplicates.
}
\examples{
\dontrun{
gwas_list <- list(gwas1, gwas2, gwas3)
merged_inst <- merge_exp_inst(gwas_list)
}
}
